--- front/src/app/detailContent/DetailContent.jsx
@@ -1,19 +1,0 @@
-'use client'
-
-const DetailContent = () => {
-    return (
-        <div>
-            
-            <h3>Aqui mostraremos los datos del detail</h3>
-
-        </div>
-    )
-}
-
-export default DetailContent;
-
-/*
-const detail = (props) => {
-    const {id} = props.params
-}
-*/
