+++ front/src/app/filters/layout.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function layout({children}) {
+  return (
+    <div>
+
+        <h1>filters</h1>
+        {children}
+        
+        </div>
+  )
+}
