+++ front/src/app/layout.js
@@ -10,7 +10,9 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <UserProvider>
-        <body>{children}</body>
+        <body>
+          {children}
+          </body>
       </UserProvider>
     </html>
   );
