+++ front/src/components/carousel/Carousel.module.css
